refactor: 예외 처리 클래스 및 핸들링 수정
