+++ src/main/java/woowacourse/auth/exception/InvalidLoginException.java
@@ -1,0 +1,10 @@
+package woowacourse.auth.exception;
+
+import woowacourse.exception.UnauthorizedException;
+
+public class InvalidLoginException extends UnauthorizedException {
+
+    public InvalidLoginException(String message) {
+        super(message);
+    }
+}
