+++ src/main/java/woowacourse/exception/UnauthorizedException.java
@@ -1,0 +1,8 @@
+package woowacourse.exception;
+
+public class UnauthorizedException extends RuntimeException {
+
+    public UnauthorizedException(final String message) {
+        super(message);
+    }
+}
