+++ src/main/java/woowacourse/order/exception/InvalidOrderException.java
@@ -1,6 +1,8 @@
 package woowacourse.order.exception;
 
-public class InvalidOrderException extends RuntimeException {
+import woowacourse.exception.BadRequestException;
+
+public class InvalidOrderException extends BadRequestException {
 
     public InvalidOrderException() {
         this("유효하지 않은 주문입니다.");
