+++ src/main/java/woowacourse/product/exception/InvalidProductException.java
@@ -1,6 +1,9 @@
-package woowacourse.shoppingcart.exception;
+package woowacourse.product.exception;
 
-public class InvalidProductException extends RuntimeException {
+import woowacourse.exception.BadRequestException;
+
+public class InvalidProductException extends BadRequestException {
+
     public InvalidProductException() {
         this("올바르지 않은 사용자 이름이거나 상품 아이디 입니다.");
     }
