--- src/main/java/woowacourse/shoppingcart/exception/EmptyResultException.java
@@ -1,7 +1,0 @@
-package woowacourse.shoppingcart.exception;
-
-public class EmptyResultException extends RuntimeException {
-    public EmptyResultException(String message) {
-        super(message);
-    }
-}
