--- src/main/java/woowacourse/shoppingcart/exception/LoginException.java
@@ -1,7 +1,0 @@
-package woowacourse.shoppingcart.exception;
-
-public class LoginException extends RuntimeException {
-    public LoginException(String message) {
-        super(message);
-    }
-}
