+++ src/test/java/woowacourse/product/domain/ProductTest.java
@@ -7,7 +7,7 @@ import static org.junit.jupiter.api.Assertions.assertAll;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-import woowacourse.shoppingcart.exception.InvalidProductException;
+import woowacourse.product.exception.InvalidProductException;
 
 public class ProductTest {
